// src/journal/dto/create-journal.dto.ts
export class CreateJournalDto {
    sources_of_hope : string;
    sources_of_hope_image: string;
    red_flags:   string;
    red_flags_image: string;
    self_care_strategies:   string;
    self_care_strategies_image: string;
    helpful_activites:   string;
    helpful_activites_image: string;
    date:   string;
}